Addressed a review comment
