+++ libcxx/test/std/utilities/smartptr/unique.ptr/unique.ptr.class/unique.ptr.observers/dereference.verify.cpp
@@ -16,9 +16,28 @@
 
 #include "test_macros.h"
 
-void f() {
+void test {
   std::unique_ptr<int[]> p(new int(3));
   const std::unique_ptr<int[]>& cp = p;
-  TEST_IGNORE_NODISCARD(*p);  // expected-error-re {{indirection requires pointer operand ('std::unique_ptr<int{{[ ]*}}[]>' invalid)}}
-  TEST_IGNORE_NODISCARD(*cp); // expected-error-re {{indirection requires pointer operand ('const std::unique_ptr<int{{[ ]*}}[]>' invalid)}}
+  TEST_IGNORE_NODISCARD(
+      *p); // expected-error-re {{indirection requires pointer operand ('std::unique_ptr<int{{[ ]*}}[]>' invalid)}}
+  TEST_IGNORE_NODISCARD(
+      *cp); // expected-error-re {{indirection requires pointer operand ('const std::unique_ptr<int{{[ ]*}}[]>' invalid)}}
+}
+
+void test_lwg4196 {
+  struct Deleter {
+    using pointer = long*;
+
+    void operator()(pointer) const {}
+  };
+
+  long l = 0;
+  std::unique_ptr<const int, Deleter> p{&l};
+
+  // expected-error-re@*:* {{static assertion failed due to requirement {{.+}}The returned reference must not bind to a temporary object.}}
+#if TEST_STD_VER >= 26
+  // expected-error@*:* {{returning reference to local temporary object}}
+#endif
+  [[maybe_unused]] int i = *p;
 }
